refactoring Segments types
