+++ ui/.eslintrc.js
@@ -9,15 +9,12 @@ module.exports = {
     extends: [
         "plugin:vue/vue3-recommended",
         "eslint:recommended",
-        "@vue/typescript/recommended",
+        "@vue/typescript/recommended"
     ],
     parserOptions: {
         ecmaVersion: 2021
     },
     rules: {
-        // semi: ["warn", "never"],
-        // quotes: ["warn", "single"],
-
         "no-console": process.env.NODE_ENV === "production" ? "warn" : "off",
         "no-debugger": process.env.NODE_ENV === "production" ? "warn" : "off",
         "@typescript-eslint/interface-name-prefix": "off",
