+++ ui/index.html
@@ -5,7 +5,7 @@
         <link rel="icon" href="/favicon.ico" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <script src="https://kit.fontawesome.com/3d13b7159a.js" crossorigin="anonymous"></script>
-        <title>Optiprism Ui</title>
+        <title>Optiprism</title>
     </head>
 
     <body>
