+++ ui/src/components/events/Segments/Segments.d.ts
